Add files via upload
There are 54 stores and 33 product families in the data. The time series starts from 2013-01-01 and finishes in 2017-08-31. The dates in the test data are for the 15 days after the last date in the training data. Date range in the test data will be very important to us while we are defining a cross-validation strategy and creating new features.

Our main mission in this project is, predicting sales for each product family and store combinations.

There are 6 data that we will study on them step by step.

Train
Test
Store
Transactions
Holidays and Events
Daily Oil Price